crypto_bot/pyproject.toml
[tool.poetry]
name = "crypto_bot"
version = "0.1.0"
description = "…"
authors = ["CryptoExplainer <cryptoexplainer9@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10"
mypy = "0.910"
numpy = "1.22.3"
black = "22.3.0"
isort = "5.13.2"
python-binance = "1.0.22"
flake8 = "7.1.1"